neo4j_example/neo4j_python/requirements.txt
django==3.2
gunicorn==20.1.0
neo4j-driver==4.1.1
neomodel==4.0.2
django_neomodel==0.0.6